+++ src/sagemaker_containers/_transformer.py
@@ -12,9 +12,10 @@
 # language governing permissions and limitations under the License.
 from __future__ import absolute_import
 
+import json
 import textwrap
 
-from sagemaker_containers import _encoders, _env, _errors, _functions, _worker
+from sagemaker_containers import _encoders, _env, _errors, _functions, _status_codes, _worker
 
 
 def default_model_fn(model_dir):
@@ -150,12 +151,23 @@ class Transformer(object):
         """
         request = _worker.Request()
 
-        data = self._input_fn(request.content, request.content_type)
-        prediction = self._predict_fn(data, self._model)
-        result = self._output_fn(prediction, request.accept)
+        try:
+            data = self._input_fn(request.content, request.content_type)
+            prediction = self._predict_fn(data, self._model)
+            result = self._output_fn(prediction, request.accept)
+        except _errors.ClientError as e:
+            return self._error_response(e, _status_codes.BAD_REQUEST)
+        except _errors.UnsupportedContentTypeError as e:
+            return self._error_response(e, _status_codes.UNSUPPORTED_MEDIA_TYPE)
+        except _errors.UnsupportedAcceptTypeError as e:
+            return self._error_response(e, _status_codes.NOT_ACCEPTABLE)
 
         if isinstance(result, tuple):
             # transforms tuple in Response for backwards compatibility
             return _worker.Response(response=result[0], accept=result[1])
 
         return result
+
+    def _error_response(self, error, status_code):
+        body = json.dumps({'error': str(error)})
+        return _worker.Response(response=body, status=status_code)