+++ test/unit/test_transfomer.py
@@ -10,6 +10,8 @@
 # distributed on an 'AS IS' BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
+import json
+
 from mock import MagicMock, patch
 import pytest
 
@@ -69,12 +71,31 @@ def test_transformer_initialize_with_client_error():
 ])
 @patch('sagemaker_containers._worker.Request', lambda: request)
 def test_transformer_transform_with_client_error(input_fn, predict_fn, output_fn):
-    with pytest.raises(_errors.ClientError) as e:
-        transform = _transformer.Transformer(model_fn=MagicMock(), input_fn=input_fn,
-                                             predict_fn=predict_fn, output_fn=output_fn)
+    transform = _transformer.Transformer(model_fn=MagicMock(), input_fn=input_fn,
+                                         predict_fn=predict_fn, output_fn=output_fn)
 
-        transform.transform()
-    assert e.value.args[0] == error_from_fn
+    response = transform.transform()
+    assert response.status_code == _status_codes.BAD_REQUEST
+    assert json.loads(response.response[0].decode('utf-8'))['error'] == str(error_from_fn)
+
+
+def test_transformer_transform_with_unsupported_content_type():
+    bad_request = test.request(data=None, content_type='fake/content-type')
+    with patch('sagemaker_containers._worker.Request', lambda: bad_request):
+        response = _transformer.Transformer().transform()
+
+    assert response.status_code == _status_codes.UNSUPPORTED_MEDIA_TYPE
+
+
+def test_transformer_transform_with_unsupported_accept_type():
+    def empty_fn(*args):
+        pass
+
+    bad_request = test.request(data=None, accept='fake/content_type')
+    with patch('sagemaker_containers._worker.Request', lambda: bad_request):
+        response = _transformer.Transformer(model_fn=empty_fn, input_fn=empty_fn, predict_fn=empty_fn).transform()
+
+    assert response.status_code == _status_codes.NOT_ACCEPTABLE
 
 
 @patch('sagemaker_containers._worker.Request', lambda: request)